Update View.py
Try to display all student data in classInfo. Success if not apiview but if apiview cant view because it say  "detail": "Authentication credentials were not provided.". I look at the internet and it say something with providing token. I thought when logged in already create the token. Now im stuck because i still searching for the answer on how to send the authentication detail (token) even already sign in.